#76 [Feat] Category Entity 삭제, 해당 객체를 enum으로 변환

+++ src/main/java/com/sparta/soomtut/lecture/controller/LectureController.java
@@ -68,7 +68,7 @@ public class LectureController {
     // 수업 전체 조회
     @GetMapping
     public ResponseEntity<?> getAllLeuctures(
-        @RequestParam(required = false, value = "category") Long category,
+        @RequestParam(required = false, value = "category") int category,
         @ModelAttribute PageRequestDto pageable
     ){
         var data = boardService.getAllPost(category, pageable.toPageable());
